Move some tests and make the code look better'
